requirements.txt
psycopg2-binary
beautifulsoup4
selenium==4.4.3
starlette